+++ DataAccessLayer/Models/Order/TOrderItems.cs
@@ -8,13 +8,13 @@ public class TOrderItems
     public int UserOrderId { get; set; }
     public virtual TProducts? Product { get; set; }
     public int ProductId { get; set; }
-    public int PurchasePrice { get; set; }
-    public int SellingPrice { get; set; }
+    public decimal PurchasePrice { get; set; }
+    public decimal SellingPrice { get; set; }
     public int Quantity { get; set; }
 
-    public int Amount { get; set; }
+    public decimal Amount { get; set; }
 
-    public int Profit {  get; set; }
+    public decimal Profit {  get; set; }
 
     [Timestamp]
     public byte[]? RowVersion { get; set; }
